faster and better: png or jpeg, crop.
if source file is a jpeg, then speed up and generate thumbnails in
format "JPEG" that are smaller and faster to load.

also define the aspect to be cropped rather than keep aspect as we're
showing the images in a fixed square. The aspect one would be good if
we did not use a grid, rather a flow layout.



SVN revision: 53218