one more day of work on ephoto, still not complete.
flow browser (why that name?!) is back, with the my traditional
combination of 255,0 in rgb colored buttons in edje.

tomorrow will work on slideshow and make it work.

NOTE: ProFUSION team will work on a solution to have elm_toolbar +
elm_menu to be good and provide what we have in eve/enjoy in a more
generic way without thousand lines of macros in EDC. The goal is to
have configurable implementation that would provide some UI on desktop
and another on embedded, also showing buttons as we have space, and
hidding them with an automatic "more" entry if there is no space for
all items. That should improve eve, enjoy, ephoto and possibly others.



SVN revision: 53250