ephoto is mostly ready again!
Well, at least that was what I had time and stamina to do. Please join
the development with your ideas and requirements.

The toolbar is now placed on top, because default.edc (from
elementary) says so. We better use elementary's profile to define the
theme to use, and a theme could have it at the bottom (mobiles?).

I've added all options to toolbar, which makes it very wide on single
image viewer. We'll later introduce toolbar item "visibility priority"
to auto-hide items and add a "More" menu if the elementary's profile
says so, other options will be to make it scrollable (as it is now).

EXIF rotation support is not right in my opinion. We're using map to
rotate, but it is not doing fine, as I said in another bug. We should
have code to request evas to load the image rotated, then we re-load
the image and it should work better (need to check how that would work
with region load from JPEG, used by elm_photocam).

Preferences are out as I ran out of time.



SVN revision: 53319