damn async things, they're always more complex than what you want.
in few cases we could quickly change directories and views would still
be populating, then they could access dt->entry
(ephoto_directory_thumb.c), so we need notifiers to know it is gone
and not mess with it any further.

Then ephoto_entry_free_listener_add() is born, added it to
flow_browser and slideshow for the sake of it... it shouldn't be
needed.



SVN revision: 53341