Ephoto: Use standard icons, remove unused code, and match dark theme.
Use standard icons based on icon theme to remove dependency on Faenza icons.  Removed the old directory code that would provide a preview of the images in the directory.  This code hasn't worked in a while and frankly is just unneccasry and would slow down the app. Remove images and theme that was shipped with Ephoto.  It is no longer needed.  Change style on some widgets to better match dark theme. Prepare Ephoto to be released soon.